Implement task-local exclusive data API in OSTD
